Faster frailty reest trick
Switching to offset trick, to speed up reestimation of demeaned model w/frailty.  If the SEs are relevant, though, the offset trick will yield the wrong SE vals.  As a consequence, also added new option ("seyes") for future functionality, in case there's ever an instance where we need the correct SEs.  (At present, the coefficient SEs are completely irrelevant for any of the calcs.)