Rework rand name gen, tweak printed msgs
In certain situations, msttvc advanced the RNG seed behind the scenes to try generating a unique int term var name.  That's dangerous, if user's unaware of that behavior AND the RNG state's relevant to other parts of user's code.

Unrelated to above, also tweaked some of the cmd's printed messages.