Fixes `covarDemean` varname lists, more tidy funct fixes
Was iterating over all names appearing in either equation.  Needed to be subsetted to the names in *each* equation to prevent a possible error in situations where a covariate's main effects aren't collapsed, but the TVEs are.  Works because TICs and TVCs were already being addressed separately.

(Causes a problem b/c `covarDemean` uses collinearity checks to figure out the relevant variables for each transition.  If main + tvc in one varlist (`namesB`), the separate effects will be collinear with the collapsed effect for the effect-collapsed transitions, meaning one of the separate effects might not get classified as being 'relevant' for that transition, causing downstream problems.)