Capturing second place where insuff obs error can get thrown
Previous place you were catching the error (after stratum-specific stcox) presupposed there were enough observations to estimate the model, but not enough for the PH test.  That overlooked a very obvious situation, in retrospect: one in which there aren't enough obsvs in the stratum to even estimate the stratum-specific stcox.