Prevent FYI msgs from printing if output suppressed
The messages in question will print by default, even if wrapped in `quietly`s (or from `run`), b/c anything displayed as an error (or in red) is force-printed by Stata.  The only way to suppress such messages (in this case, non-error messages that happen to be printed in the error msg color) is to `capture` them.